// packages/satteri/src/mdx-types.ts
export interface MdxJsxAttributeValueExpression extends Node {
  type: "mdxJsxAttributeValueExpression";
  value: string;
  data?: MdxJsxAttributeValueExpressionData | undefined;
}

export interface MdxJsxAttributeValueExpressionData extends Data {}

export interface MdxJsxExpressionAttribute extends Node {
  type: "mdxJsxExpressionAttribute";
  value: string;
  data?: MdxJsxExpressionAttributeData | undefined;
}

export interface MdxJsxExpressionAttributeData extends Data {}

export interface MdxJsxAttribute extends Node {
  type: "mdxJsxAttribute";
  name: string;
  value?: MdxJsxAttributeValueExpression | string | null | undefined;
  data?: MdxJsxAttributeData | undefined;
}

export interface MdxJsxAttributeData extends Data {}

export interface MdxJsxFlowElement extends MdastParent {
  type: "mdxJsxFlowElement";
  name: string | null;
  attributes: Array<MdxJsxAttribute | MdxJsxExpressionAttribute>;
  children: Array<BlockContent | DefinitionContent>;
  data?: MdxJsxFlowElementData | undefined;
}

export interface MdxJsxFlowElementData extends MdastData {}

export interface MdxJsxTextElement extends MdastParent {
  type: "mdxJsxTextElement";
  name: string | null;
  attributes: Array<MdxJsxAttribute | MdxJsxExpressionAttribute>;
  children: PhrasingContent[];
  data?: MdxJsxTextElementData | undefined;
}

export interface MdxJsxTextElementData extends MdastData {}
// ...
export interface MdxFlowExpression extends MdastLiteral {
  type: "mdxFlowExpression";
  data?: MdxFlowExpressionData | undefined;
}

export interface MdxFlowExpressionData extends MdastData {}

export interface MdxTextExpression extends MdastLiteral {
  type: "mdxTextExpression";
  data?: MdxTextExpressionData | undefined;
}

export interface MdxTextExpressionData extends MdastData {}
// ...
export interface MdxjsEsm extends MdastLiteral {
  type: "mdxjsEsm";
  data?: MdxjsEsmData | undefined;
}

export interface MdxjsEsmData extends MdastData {}